charito-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
requests==2.31.0
paho-mqtt==2.1.0
certifi>=2024.2.2
-e ../shared/timeauthority-pkg
